requirements.txt
Django==4.2.19
mysqlclient==2.2.7

django-crispy-forms==2.3
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.10
django-select2==8.2.3
django-autocomplete-light==3.11.0
django-ckeditor==6.7.2

django-countries==7.6.1
django-money==3.5.3
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-xhtml2pdf==0.0.4

xhtml2pdf==0.2.16
reportlab==4.3.1
pyHanko==0.25.3
pypdf==5.3.0

Pillow==11.1.0
openpyxl==3.1.5
qrcode==8.0

babel==2.17.0
arabic-reshaper==3.0.0
python-bidi==0.6.3

requests==2.32.3
